requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0
Flask-Bcrypt>=1.0
Flask-JWT-Extended>=4.0
psycopg2-binary # For PostgreSQL, if you use it. For SQLite, this is not strictly needed but good to have.
python-dotenv>=0.15 # For managing environment variables from .env file
flasgger>=0.9.5 # For OpenAPI/Swagger documentation
